fix: o dataclass do evento sai de providers para inbound (gate estrutural) [#36]
test_no_direct_provider_imports barra modulo de dominio importando
app.core.whatsapp.providers.*; TemplateStatusEvent passa a morar em
core/whatsapp/inbound.py, ao lado de InboundMessage. E a 0079 muda a
forma de public_whatsapp_accounts, entao o teste de colunas acompanha.
